Cleaned up App.js, installed bootstrap and organised the files
